Guild implementation type with new create2deployer and guildInitialized events
