+++ package.json
@@ -9,11 +9,12 @@
     "postinstall": "npm run codegen",
     "codegen": "node scripts/generate-schema.js && node scripts/generate-subgraph.js && graph codegen --output-dir src/types",
     "build": "graph build",
-    "start-local": "yarn create-local && yarn deploy-local",
+    "start-local": "yarn run updateBytecodes && yarn create-local && yarn deploy-local",
     "create-local": "graph create mprasanjith/dxdao --node http://127.0.0.1:8020",
     "deploy-local": "graph deploy mprasanjith/dxdao --ipfs http://127.0.0.1:5001 --node http://127.0.0.1:8020 -l v0.0.1",
     "deploy": "graph deploy mprasanjith/dxdao --ipfs https://api.thegraph.com/ipfs/ --node https://api.thegraph.com/deploy/",
     "clean": "rm -rf ./data && docker compose down",
+    "updateBytecodes": "node ./scripts/updateBytecodes.js",
     "dev": "./scripts/dev.sh",
     "start-docker": "pnpm run clean && docker compose up"
   },
@@ -30,6 +31,8 @@
   "dependencies": {
     "babel-polyfill": "^6.26.0",
     "babel-register": "^6.26.0",
-    "js-yaml": "^4.1.0"
+    "js-yaml": "^4.1.0",
+    "dxdao-contracts": "*"
   }
 }
+
