+++ src/mappings/Create2Deployer/datasource.yaml
@@ -6,7 +6,6 @@ abis:
   - SnapshotERC20Guild
 entities:
   - Guild
-  - Token
 eventHandlers:
   - event: Deployed(address,bytes32)
     handler: handleDeployedEvent
