+++ src/mappings/Create2Deployer/local.ts
@@ -1,0 +1,1 @@
+export const local = "[\n  {\n    \"type\": \"ERC20Guild\",\n    \"bytecodeHash\": \"0xce75958407594ab8566c529b45d36016a560676f5869195eb8adebc4987a3757\",\n    \"deployedBytecodeHash\": \"0xbad237e9eb436605b0504edd53755d1bf0e013c8a81d3126bed7001a9de96e02\",\n    \"features\": []\n  },\n  {\n    \"type\": \"SnapshotRepERC20Guild\",\n    \"bytecodeHash\": \"0x06f9cae443d8d0b056fa744147dbbbbc3da47996ff10b6272096581231e642cf\",\n    \"deployedBytecodeHash\": \"0xf29a894895da85cf317fb94bcd56a2a066d505c0935f3a9590ae8410dcacd5a4\",\n    \"features\": [\n      \"REP\",\n      \"SNAPSHOT\"\n    ]\n  },\n  {\n    \"type\": \"SnapshotERC20Guild\",\n    \"bytecodeHash\": \"0xb8c656709061554709c9feb9cfcc3ae9df8e86a65af0bf74c55ef0e66e17818d\",\n    \"deployedBytecodeHash\": \"0xb21342cc31da5670be91a4422b9c2c70764d93422880bc9ebb8cdd7e629eb6b3\",\n    \"features\": [\n      \"SNAPSHOT\"\n    ]\n  },\n  {\n    \"type\": \"DXDGuild\",\n    \"bytecodeHash\": \"0xd8382264cf76dcc4ac93b99a30e56084a87000383d2e15739021776008cd566c\",\n    \"deployedBytecodeHash\": \"0x9fd9e08f4c1654ea8791d74ca244e77fb2825d38f473eb8566658b1272605a9e\",\n    \"features\": []\n  }\n]";
